Move SpinnerPage to containers
